+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/core.module.ts
@@ -1,7 +1,7 @@
 import {AndroidContentProviderPlugin} from './platform-plugins/cordova-plugins/android-content-provider-plugin';
 import {BrowserPrinterPlugin} from './platform-plugins/printers/browser-printer.plugin';
 import {PRINTERS} from './platform-plugins/printers/printer.service';
-import {ConsoleScannerPlugin} from './platform-plugins/image-scanners/console-scanner/console-scanner.plugin';
+import {ConsoleScannerPlugin} from './platform-plugins/barcode-scanners/console-scanner/console-scanner.plugin';
 import { SessionService } from './services/session.service';
 import { PersonalizationStartupTask } from './startup/personalization-startup-task';
 import { STARTUP_TASKS, STARTUP_FAILED_COMPONENT } from './services/startup.service';
@@ -65,17 +65,17 @@ import { AudioRepositoryService } from './audio/audio-repository.service';
 import { AudioInteractionService } from './audio/audio-interaction.service';
 import { AudioConsolePlugin } from './audio/audio-console.plugin';
 import { CapacitorStatusBarPlatformPlugin } from './startup/capacitor-status-bar-platform-plugin';
-import { ScanditCapacitorImageScanner } from './platform-plugins/image-scanners/scandit-capacitor/scandit-capacitor.service';
-import { IMAGE_SCANNERS, SCANNERS } from './platform-plugins/image-scanners/scanner';
-import { BarcodeScanner } from './platform-plugins/image-scanners/barcode-scanner.service';
+import { ScanditCapacitorImageScanner } from './platform-plugins/barcode-scanners/scandit-capacitor/scandit-capacitor.service';
+import { IMAGE_SCANNERS, SCANNERS } from './platform-plugins/barcode-scanners/scanner';
+import { BarcodeScanner } from './platform-plugins/barcode-scanners/barcode-scanner.service';
 import {ClientExecutableService} from "./services/client-executable.service";
 import { CapacitorIosPlatform } from './platforms/capacitor-ios.platform';
 import { CapacitorAndroidPlatform } from './platforms/capacitor-android.platform';
-import { AilaScannerCordovaPlugin } from './platform-plugins/image-scanners/aila-scanner-cordova/aila-scanner-cordova.plugin';
-import { InfineaScannerCordovaPlugin } from './platform-plugins/image-scanners/infinea-scanner-cordova/infinea-scanner-cordova.plugin';
-import { WedgeScannerPlugin } from './platform-plugins/image-scanners/wedge-scanner/wedge-scanner.plugin';
-import { ServerScannerPlugin } from './platform-plugins/image-scanners/server-scanner/server-scanner.service';
-import { ScanditScannerCordovaPlugin } from './platform-plugins/image-scanners/scandit-scanner-cordova/scandit-scanner-cordova.plugin';
+import { AilaScannerCordovaPlugin } from './platform-plugins/barcode-scanners/aila-scanner-cordova/aila-scanner-cordova.plugin';
+import { InfineaScannerCordovaPlugin } from './platform-plugins/barcode-scanners/infinea-scanner-cordova/infinea-scanner-cordova.plugin';
+import { WedgeScannerPlugin } from './platform-plugins/barcode-scanners/wedge-scanner/wedge-scanner.plugin';
+import { ServerScannerPlugin } from './platform-plugins/barcode-scanners/server-scanner/server-scanner.service';
+import { ScanditScannerCordovaPlugin } from './platform-plugins/barcode-scanners/scandit-scanner-cordova/scandit-scanner-cordova.plugin';
 
 registerLocaleData(locale_enCA, 'en-CA');
 registerLocaleData(locale_frCA, 'fr-CA');
