+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/barcode-scanners/aila-scanner-cordova/aila-to-openpos-barcode-type.ts
