+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/barcode-scanners/infinea-scanner-cordova/infinea-scanner-cordova.plugin.ts
