+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/barcode-scanners/scandit-capacitor/scandit-capacitor.service.ts
