+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/core/platform-plugins/barcode-scanners/scandit-scanner-cordova/scandit-to-openpos-barcode-type.ts
