+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/shared/screen-parts/scan-or-search/scan-or-search.component.ts
@@ -17,8 +17,8 @@ import {MediaBreakpoints, OpenposMediaService} from '../../../core/media/openpos
 import {Observable, Subject, Subscription} from 'rxjs';
 import {OnBecomingActive} from '../../../core/life-cycle-interfaces/becoming-active.interface';
 import {OnLeavingActive} from '../../../core/life-cycle-interfaces/leaving-active.interface';
-import { BarcodeScanner } from '../../../core/platform-plugins/image-scanners/barcode-scanner.service';
-import { ScanData } from '../../../core/platform-plugins/image-scanners/scanner';
+import { BarcodeScanner } from '../../../core/platform-plugins/barcode-scanners/barcode-scanner.service';
+import { ScanData } from '../../../core/platform-plugins/barcode-scanners/scanner';
 
 @ScreenPart({
     name: 'scanOrSearch'
