+++ openpos-client-libs/projects/openpos-client-core-lib/src/lib/shared/shared.module.ts
@@ -166,7 +166,7 @@ import { StatusDetailsItemComponent } from './status/status-details/item/status-
 import { StatusBarComponent } from './status/status-bar/status-bar.component';
 import { LocalizePipe } from './pipes/localize.pipe';
 import { ModalOverlayContainerDirective } from './modal-overlay-container/modal-overlay-container.directive';
-import { ImageScannerComponent } from '../core/platform-plugins/image-scanners/image-scanner.component';
+import { ImageScannerComponent } from '../core/platform-plugins/barcode-scanners/image-scanner.component';
 
 const screenParts = [
     TenderPartComponent,
